+++ src/components/Tasks/Create-Task/TaskDrawerMenu.tsx
@@ -1,11 +1,12 @@
 import { makeStyles } from "@material-ui/core";
-import { Divider, Grid, TextField } from "@mui/material";
+import { Divider, Grid, TextField, Typography } from "@mui/material";
 import Autocomplete from "@mui/material/Autocomplete";
 import { CBox } from "components/material-ui";
 import React, { useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 // import { AttachmentIcon } from "components/material-ui/icons";
 import assets from "assets/assets";
+import ErrorOutlineOutlinedIcon from '@mui/icons-material/ErrorOutlineOutlined';
 import CButton from "components/Button/Button";
 import CDatePicker from "components/DatePicker/CDatePicker";
 import CustomModal from "components/Modal";
@@ -35,6 +36,8 @@ import { RootState } from "redux/reducers";
 import ViewAllDocs from "../SubTasks/ViewAllDocs";
 import docsAction from "redux/action/docs.actions";
 import { DOCS_CONFIG } from "config/docs.config";
+import { useConfirm } from "material-ui-confirm";
+import { CustomStack } from "components/TaskComponent/Tabs/TaskCard";
 
 interface Props {
   taskMenue: TaskInterface;
@@ -47,6 +50,7 @@ function TaskDrawerMenu({ taskMenue, subtasks }: Props) {
   const [opencdrawer, setOpenCDrawer] = useState<boolean>(false);
   const classes = useStyles();
   const dispatch = useDispatch();
+  const confirm = useConfirm();
 
   const {
     _id,
@@ -276,9 +280,46 @@ function TaskDrawerMenu({ taskMenue, subtasks }: Props) {
     descriptionInputRef.current.value = prevObj.description;
   };
 
+  // const handleDeleteTask = (e: any) => {
+  //   e.stopPropagation();
+  //   dispatch(
+  //     deleteTask({
+  //       other: _id,
+  //       success: (res) => {
+  //         dispatch({ type: TASK_CONFIG.PULL_TASK_FROM_STORE, payload: _id });
+  //         dispatch(taskActions.closeTaskDrawer());
+  //         toast.success("Task deleted");
+  //         //deleted task id = _id
+  //       },
+  //       onFailAction: () => {
+  //         toast.error("Failed to delete task!");
+  //       },
+  //     })
+  //   );
+  //   setShowUpdateBtn(false);
+  // };
   const handleDeleteTask = (e: any) => {
     e.stopPropagation();
-    dispatch(
+    confirm({
+      title: <CustomStack gap={1}><ErrorOutlineOutlinedIcon/> Confirmation</CustomStack>,
+      description:<Typography sx={{color:'#605C5C', fontSize:13, fontWeight:'500', pt:2}}>Are you sure you want to delete this task?</Typography>,
+      titleProps: { color: "red", borderBottom:'1px solid #D3D4D9' },
+      confirmationText:"Delete",
+      confirmationButtonProps: {sx:{textTransform:'capitalize'}, variant:"outlined", color:"error"},
+      cancellationText: <CButton
+      variant="contained"
+      elevation={0}
+      styles={{
+        color: "#605C5C",
+        backgroundColor: "#ECF0F1",
+        fontSize: 12,
+        fontWeight: "bold",
+      }}
+      label={"Cancel"}
+    />,
+
+    }).then(() => {
+          dispatch(
       deleteTask({
         other: _id,
         success: (res) => {
@@ -293,8 +334,8 @@ function TaskDrawerMenu({ taskMenue, subtasks }: Props) {
       })
     );
     setShowUpdateBtn(false);
+    });
   };
-
   const handleProjectChange = (project: any) => {
     if (project === null) {
       setAdminListOpt([]);
